refactor: icons folder
