+++ package.json
@@ -23,7 +23,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "fmt": "prettier --write ./src/**"
+    "fmt": "prettier --write ./src/**",
+    "check": "tsc --noEmit"
   },
   "eslintConfig": {
     "extends": [
