+++ src/components/Cart.tsx
@@ -3,7 +3,7 @@ import Card from "./Card";
 import { Context } from "./Context";
 import { useContext, useEffect } from "react";
 import { Link } from "react-router-dom";
-import { RightChevrons } from "./icons";
+import RightChevrons from "../icons/RightChevrons";
 
 export default function Cart() {
   const { state, actions } = useContext(Context)!;
