+++ src/icons/props.ts
@@ -1,0 +1,4 @@
+export default interface IconProps {
+  size: number;
+  className?: string;
+}
